Update Challenge 3 NOT WORKING.py
Got it working!

Code section stating 163 - Instead of editing the array `self.players` directly in the for loop, I created a new array (`sp`) and added the ones needed to. Once completed, I equated `self.players` to `sp` to get the correct list of objects as `self.players`.

Also, brought `again` condition for `while` loop into `game()` function so that if there are no players left in `self.players`, the game would end.